fix: update test case
