+++ package.json
@@ -12,7 +12,7 @@
     "@changesets/cli": "^2.18.1",
     "@commitlint/cli": "^15.0.0",
     "@commitlint/config-conventional": "^15.0.0",
-    "@emotion/jest": "^11.6.0",
+    "@emotion/jest": "^11.7.1",
     "@storybook/addon-a11y": "^6.4.9",
     "@storybook/addon-actions": "^6.4.9",
     "@storybook/addon-essentials": "^6.4.9",
@@ -24,17 +24,16 @@
     "@storybook/theming": "^6.4.9",
     "@testing-library/jest-dom": "^5.15.1",
     "@testing-library/react": "^12.1.2",
-    "@types/jest": "^26.0.2",
+    "@types/jest": "^26.0.24",
     "babel-loader": "^8.2.3",
     "husky": "^7.0.4",
     "jest": "^26.3.1",
     "jest-electron": "^0.1.12",
     "jest-environment-jsdom": "^26.3.1",
-    "jest-watch-typeahead": "^1.0.0",
     "lerna": "^4.0.0",
     "prettier": "^2.5.1",
     "react-icons": "^4.3.1",
-    "ts-jest": "^26.0.7",
+    "ts-jest": "^26.5.6",
     "zx": "^4.2.0"
   },
   "workspaces": [
