+++ packages/typography/tests/paragraph.test.tsx
@@ -28,7 +28,7 @@ test("Paragraph renders with indent", () => {
       graphic design) may also be considered to be a design activity.</Paragraph>
   </Typography>)
   expect(screen.getByTestId("test-paragraph")).toHaveStyle({
-    "text-indent": "32px",
+    "text-indent": "2em",
   })
 })
 
@@ -74,44 +74,5 @@ test("Paragraph renders with disabled style", () => {
   expect(screen.getByText("This is a paragraph")).toHaveStyle({
     "cursor": "not-allowed",
   })
-})
-
-test("Paragraph renders short text with ellipsis", () => {
-  render(<Typography>
-    <Paragraph ellipsis={true}>
-      This is a paragraph
-    </Paragraph>
-  </Typography>)
-  expect(screen.queryByText("...")).not.toBeInTheDocument()
-})
-
-test("Paragraph render with ellipsis", () => {
-  const onExpand = jest.fn()
-  render(<Typography>
-    <Paragraph data-testid="test-paragraph"
-               ellipsis={new EllipsisBuilder()
-                 .rows(2)
-                 .expandable(true)
-                 .expandLabel("TestExpand")
-                 .onExpand(onExpand)
-                 .suffix("TestSuffix")
-                 .create()}>
-      A design is a plan or
-      specification for the
-      construction of
-      an object or system or for
-      the
-      implementation of an activity or process, or the result of that plan or specification in the form of a prototype,
-      product or process. The verb to design expresses the process of developing a design. In some cases, the direct
-      construction of an object without an explicit prior plan (such as in craft work, some engineering, coding, and
-      graphic design) may also be considered to be a design activity.
-    </Paragraph>
-  </Typography>)
-  expect(screen.getByText("TestSuffix")).toBeInTheDocument()
-  expect(screen.getByText("...")).toBeInTheDocument()
-  expect(screen.getByText("TestExpand")).toBeInTheDocument()
-  fireEvent.click(screen.getByText("TestExpand"))
-  expect(onExpand).toBeCalled()
-  expect(screen.queryByText("...")).not.toBeInTheDocument()
 })
